+++ exportvector.cpp
@@ -277,7 +277,7 @@ void PdfFileWriter::StartFile(void) {
 "%%%c%c%c%c\r\n",
         0xe2, 0xe3, 0xcf, 0xd3);
     
-    xref[1] = ftell(f);
+    xref[1] = (DWORD)ftell(f);
     fprintf(f,
 "1 0 obj\r\n"
 "  << /Type /Catalog\r\n"
@@ -286,7 +286,7 @@ void PdfFileWriter::StartFile(void) {
 "  >>\r\n"
 "endobj\r\n");
 
-    xref[2] = ftell(f);
+    xref[2] = (DWORD)ftell(f);
     fprintf(f,
 "2 0 obj\r\n"
 "  << /Type /Outlines\r\n"
@@ -294,7 +294,7 @@ void PdfFileWriter::StartFile(void) {
 "  >>\r\n"
 "endobj\r\n");
 
-    xref[3] = ftell(f);
+    xref[3] = (DWORD)ftell(f);
     fprintf(f,
 "3 0 obj\r\n"
 "  << /Type /Pages\r\n"
@@ -303,7 +303,7 @@ void PdfFileWriter::StartFile(void) {
 "  >>\r\n"
 "endobj\r\n");
 
-    xref[4] = ftell(f);
+    xref[4] = (DWORD)ftell(f);
     fprintf(f,
 "4 0 obj\r\n"
 "  << /Type /Page\r\n"
@@ -318,35 +318,35 @@ void PdfFileWriter::StartFile(void) {
             MmToPts(ptMax.x - ptMin.x),
             MmToPts(ptMax.y - ptMin.y));
 
-    xref[5] = ftell(f);
+    xref[5] = (DWORD)ftell(f);
     fprintf(f,
 "5 0 obj\r\n"
 "  << /Length 6 0 R >>\r\n"
 "stream\r\n");
-    bodyStart = ftell(f);
+    bodyStart = (DWORD)ftell(f);
 }
 
 void PdfFileWriter::FinishAndCloseFile(void) {
-    DWORD bodyEnd = ftell(f);
+    DWORD bodyEnd = (DWORD)ftell(f);
 
     fprintf(f,
 "endstream\r\n"
 "endobj\r\n");
 
-    xref[6] = ftell(f);
+    xref[6] = (DWORD)ftell(f);
     fprintf(f,
 "6 0 obj\r\n"
 "  %d\r\n"
 "endobj\r\n",
         bodyEnd - bodyStart);
 
-    xref[7] = ftell(f);
+    xref[7] = (DWORD)ftell(f);
     fprintf(f,
 "7 0 obj\r\n"
 "  [/PDF /Text]\r\n"
 "endobj\r\n");
 
-    xref[8] = ftell(f);
+    xref[8] = (DWORD)ftell(f);
     fprintf(f,
 "8 0 obj\r\n"
 "  << /Type /Font\r\n"
@@ -357,13 +357,13 @@ void PdfFileWriter::FinishAndCloseFile(void) {
 "  >>\r\n"
 "endobj\r\n");
 
-    xref[9] = ftell(f);
+    xref[9] = (DWORD)ftell(f);
     fprintf(f,
 "9 0 obj\r\n"
 "  << /Creator (SolveSpace)\r\n"
 "  >>\r\n");
     
-    DWORD xrefStart = ftell(f);
+    DWORD xrefStart = (DWORD)ftell(f);
     fprintf(f,
 "xref\r\n"
 "0 10\r\n"
